V01

+++ R/stream.R
@@ -44,6 +44,14 @@
 #' @param file_name Character with name of file. By default, a
 #'   temporary file is created, tweets are parsed and returned to
 #'   parent environment, and the temporary file is deleted.
+#' @param write_json_stdout Redirects the data stream to the standard output. 
+#'   By default, \code{write_json_stdout = FALSE}. This approach
+#'   is useful for piping the output stream, which allows the processing
+#'   of completed blocks of data while maintaining the data stream open to 
+#'   receive new ones. For instance, the output of an R script that uses 
+#'   \code{rtweet::stream_tweets} with the \code{write_json_stdout = TRUE} option
+#'   can be fed to the \code{split} command to produce multiple files: 
+#'   \code{Rscript my_script.R | split -l1000 -d --additional-suffix .json - myTweets_}.
 #' @param verbose Logical, indicating whether or not to include output
 #'   processing/retrieval messages.
 #' @param \dots Insert magical parameters, spell, or potion here. Or
@@ -136,6 +144,7 @@ stream_tweets <- function(q = "",
                           parse = TRUE,
                           token = NULL,
                           file_name = NULL,
+                          write_json_stdout = FALSE,
                           verbose = TRUE,
                           ...) {
   if ("append" %in% names(list(...))) {
@@ -184,49 +193,66 @@ stream_tweets <- function(q = "",
   if (!grepl("\\.json$", file_name)) {
     file_name <- paste0(file_name, ".json")
   }
-  if (!file.exists(file_name)) file.create(file_name)
+  if (!file.exists(file_name) & !write_json_stdout) file.create(file_name)
   if (verbose) {
     message(
       paste0("Streaming tweets for ", timeout, " seconds...")
     )
   }
   r <- NULL
-  con <- file(file_name, "wt")
-  on.exit({sh <- tryCatch(close(con), error = function(e) return(NULL),
-    warning = function(w) return(NULL))}, add = TRUE)
 
   start_time <- Sys.time()
   stop_time <- Sys.time() + timeout
   ctr <- 0
-  while (timeout > 0) {
-    r <- tryCatch(httr::POST(
-      url = url,
-      httr::config(token = token, timeout = timeout),
-      httr::write_stream(write_fun(con)),
-      httr::add_headers(Accept = "application/json"),
-      httr::add_headers(`Accept-Encoding` = "gzip, deflate")),
-      error = function(e) return(e))
-    timeout <- as.numeric(difftime(stop_time, Sys.time(), units = "secs"))
-    if (timeout > 0) {
-      ctr <- ctr + 1
-      if (ctr == 1 && verbose) message(
-        "The stream disconnected prematurely. Reconnecting...")
-      if (ctr == 2 && verbose) message("Reconnecting again...")
-      if (ctr == 5) break
-    } else if (verbose) {
-      message("Finished streaming tweets!")
+  if(!write_json_stdout){
+    con <- file(file_name, "wt")
+    on.exit({sh <- tryCatch(close(con), error = function(e) return(NULL),
+                            warning = function(w) return(NULL))}, add = TRUE)
+    while (timeout > 0) {
+      r <- tryCatch(httr::POST(
+        url = url,
+        httr::config(token = token, timeout = timeout),
+        httr::write_stream(write_fun(con)),
+        httr::add_headers(Accept = "application/json"),
+        httr::add_headers(`Accept-Encoding` = "gzip, deflate")),
+        error = function(e) return(e))
+      timeout <- as.numeric(difftime(stop_time, Sys.time(), units = "secs"))
+      if (timeout > 0) {
+        ctr <- ctr + 1
+        if (ctr == 1 && verbose) message(
+          "The stream disconnected prematurely. Reconnecting...")
+        if (ctr == 2 && verbose) message("Reconnecting again...")
+        if (ctr == 5) break
+      } else if (verbose) {
+        message("Finished streaming tweets!")
+      }
+    }
+    close(con)
+    if (parse) {
+      out <- parse_stream(file_name, verbose = verbose)
+      if (tmp) {
+        file.remove(file_name)
+      }
+      return(out)
+    }
+    if (verbose) message("streaming data saved as ", file_name)
+    invisible(r)
+  }else{ # write_json_stdout == TRUE
+    while (timeout > 0) {
+      r <- tryCatch(httr::POST(
+        url = url,
+        httr::config(token = token, timeout = timeout),
+        httr::write_stream(function(x) {write(rawToChar(x), stdout())}),
+        httr::add_headers(Accept = "application/json"),
+        httr::add_headers(`Accept-Encoding` = "gzip, deflate")),
+        error = function(e) return(e))
+      timeout <- as.numeric(difftime(stop_time, Sys.time(), units = "secs"))
+      if (timeout > 0) {
+        ctr <- ctr + 1
+        if (ctr == 5) break
+      }
     }
   }
-  close(con)
-  if (parse) {
-    out <- parse_stream(file_name, verbose = verbose)
-    if (tmp) {
-      file.remove(file_name)
-    }
-    return(out)
-  }
-  if (verbose) message("streaming data saved as ", file_name)
-  invisible(r)
 }
 
 write_fun <- function(con) {
@@ -252,7 +278,7 @@ stream_params <- function(stream, ...) {
   op <- getOption("encoding")
   on.exit(options(encoding = op), add = TRUE)
   options(encoding = "UTF-8")
-
+  
   if (inherits(stream, "coords")) {
     stream <- stream$box
   }
@@ -337,7 +363,7 @@ stream_data <- function(file_name, ...) {
     if (length(d) > 0) {
       dd <- sapply(d, function(x) {
         o <- tryCatch(jsonlite::fromJSON(x),
-          error = function(e) return(FALSE))
+                      error = function(e) return(FALSE))
         if (identical(o, FALSE)) return(FALSE)
         return(TRUE)
       }, USE.NAMES = FALSE)
@@ -355,7 +381,7 @@ data_from_stream <- function(x, n = 10000L, n_max = -1L, ...) {
   }
   if (!requireNamespace("readr", quietly = TRUE)) {
     warning("For better performance when reading large twitter .json files, ",
-      "try installing the readr package before using this function.")
+            "try installing the readr package before using this function.")
     return(stream_data(x, ...))
   }
   ## initalize counters and output vector
@@ -393,7 +419,7 @@ data_from_stream2 <- function(x, n = 10000L, n_max = -1L, ...) {
   }
   if (!requireNamespace("readr", quietly = TRUE)) {
     warning("For better performance when reading large twitter .json files, ",
-      "try installing the readr package before using this function.")
+            "try installing the readr package before using this function.")
     return(stream_data(x, ...))
   }
   ## initalize counters and output vector
@@ -480,14 +506,14 @@ stream_tweets2 <- function(..., dir = NULL, append = FALSE) {
   if (!dir.exists(dir)) {
     new_dir(dir)
   }
-
+  
   ## capture and match dots
   dots <- match_fun(list(...), "stream_tweets")
   ## start time
   start <- Sys.time()
   ## finish time (given requested timeout)
   reqtime <- start + dots[["timeout"]]
-
+  
   ## save file name for final file
   file_name <- dots[["file_name"]]
   if (is.null(file_name)) {
@@ -501,17 +527,17 @@ stream_tweets2 <- function(..., dir = NULL, append = FALSE) {
   ## store verbose value, then override to FALSE
   verbose <- dots[["verbose"]]
   dots[["verbose"]] <- FALSE
-
+  
   ## display message if verbose
   if (verbose) {
     message(paste0("Streaming tweets for ", dots[["timeout"]], " seconds..."))
   }
-
+  
   ## initialize output vector
   rt <- list()
   ## start counter
   i <- 1L
-
+  
   ## restart and continue stream until reqtime
   while (Sys.time() <= reqtime) {
     dots[["file_name"]] <- file.path(dir, paste0(file_name, "-", i, ".json"))
@@ -578,5 +604,3 @@ stream_dir <- function() {
   timestamp <- gsub("\\s|\\:|\\-", "", substr(Sys.time(), 1, 19))
   paste0("stream-", timestamp)
 }
-
-
